requirements.txt
google-generativeai
python-dotenv
speechrecognition
pyttsx3
pyautogui
pyaudio
yt-dlp
requests
